Windows CI: test-unit on pkg\plugins
